apply review proposals
