+++ lenses-annotation-processor/src/jvmMain/kotlin/dev/fritz2/lens/LensesProcessor.kt
@@ -11,7 +11,6 @@ import com.squareup.kotlinpoet.ParameterizedTypeName.Companion.parameterizedBy
 import com.squareup.kotlinpoet.ksp.*
 import dev.fritz2.core.Lens
 import dev.fritz2.core.Lenses
-import dev.fritz2.lens.CheckResult.Companion.collectMessages
 
 /**
  * This Processor generates automatically functions of lenses for all public properties of a data class, a sealed class
@@ -146,7 +145,7 @@ private class LensesVisitor(
         classDeclaration: KSClassDeclaration,
         companionObject: KSClassDeclaration?,
         packageName: String,
-        determineLensablePropertiesStrategy: DetermineLenseableProperties,
+        determineLensablePropertiesStrategy: LenseablePropertiesDeterminer,
         checker: RequirementChecker,
         addLensCode: FunSpec.Builder.(MemberName, KSClassDeclaration) -> FunSpec.Builder
     ) {
@@ -360,11 +359,11 @@ private class LensesVisitor(
     private val createLens: FunSpec.Builder.(MemberName, KSClassDeclaration) -> FunSpec.Builder = { attributeName, _ ->
         addCode(
             """ 
-        |return %M(
-        |    "%L",
-        |    { it.%M },
-        |    { p, v -> p.copy(%M = v)}
-        |  )
+            |return %M(
+            |    "%L",
+            |    { it.%M },
+            |    { p, v -> p.copy(%M = v)}
+            |  )
         """.trimMargin(),
             MemberName("dev.fritz2.core", "lensOf"),
             attributeName,
@@ -379,11 +378,11 @@ private class LensesVisitor(
 
             addCode(
                 """ 
-            |return %M(
-            |    "%L",
-            |    { parent ->
-            |        when(parent) {
-            """.trimMargin(),
+                |return %M(
+                |    "%L",
+                |    { parent ->
+                |        when(parent) {
+                """.trimMargin(),
                 MemberName("dev.fritz2.core", "lensOf"),
                 attributeName
             )
@@ -391,285 +390,37 @@ private class LensesVisitor(
             children.forEach { child ->
                 addStatement(
                     """ 
-                |            is %T -> parent.%M
-                """.trimMargin(),
+                    |            is %T -> parent.%M
+                    """.trimMargin(),
                     child.toClassName(),
                     attributeName
                 )
             }
             addCode(
                 """
-            |        }
-            |    },
-            |    { parent, value ->
-            |        when(parent) {
-            """.trimMargin(),
+                |        }
+                |    },
+                |    { parent, value ->
+                |        when(parent) {
+                """.trimMargin(),
             )
             addStatement("")
             children.forEach { child ->
                 addStatement(
                     """ 
-                |            is %T -> parent.copy(%M = value)
-                """.trimMargin(),
+                    |            is %T -> parent.copy(%M = value)
+                    """.trimMargin(),
                     child.toClassName(),
                     attributeName
                 )
             }
             addCode(
                 """
-            |        }
-            |    }
-            |)
-            """.trimMargin(),
+                |        }
+                |    }
+                |)
+                """.trimMargin(),
             )
         }
 }
 
-private fun String.lowerCamelCased() = "${first().lowercase()}${drop(1)}"
-
-private enum class TypeVariant {
-    Invalid, DataClass, SealedDataClass, SealedInterface
-}
-
-private fun KSClassDeclaration.isTypeVariant(): TypeVariant = when (classKind) {
-    ClassKind.INTERFACE ->
-        if (modifiers.contains(Modifier.SEALED)) TypeVariant.SealedInterface
-        else TypeVariant.Invalid
-
-    ClassKind.CLASS ->
-        if (modifiers.contains(Modifier.DATA)) TypeVariant.DataClass
-        else if (modifiers.contains(Modifier.SEALED)) TypeVariant.SealedDataClass
-        else TypeVariant.Invalid
-
-    else -> TypeVariant.Invalid
-}
-
-private fun interface DetermineLenseableProperties {
-    fun determine(classDeclaration: KSClassDeclaration): List<KSPropertyDeclaration>
-}
-
-private val determineLensablePropertiesInConstructor = DetermineLenseableProperties { classDeclaration ->
-    val allPublicCtorProps = classDeclaration.primaryConstructor!!.parameters.filter { it.isVal }.map { it.name }
-    classDeclaration.getDeclaredProperties()
-        .filter { it.isPublic() && allPublicCtorProps.contains(it.simpleName) }.toList()
-}
-
-private val determineLensablePropertiesInBody = DetermineLenseableProperties { classDeclaration ->
-    classDeclaration.getDeclaredProperties().filter { it.isPublic() }.toList()
-}
-
-enum class Severity {
-    Warning, Error
-}
-
-data class Message(val severity: Severity, val message: String)
-
-fun KSPLogger.log(message: Message) {
-    when (message.severity) {
-        Severity.Warning -> this.warn(message.message)
-        Severity.Error -> this.error(message.message)
-    }
-}
-
-private sealed interface CheckResult {
-    object Success : CheckResult
-    data class Warning(val warnings: List<Message>) : CheckResult {
-        constructor(warning: Message) : this(listOf(warning))
-    }
-
-    data class Failure(val errors: List<Message>) : CheckResult {
-        constructor(error: Message) : this(listOf(error))
-    }
-
-    companion object {
-        fun collectMessages(vararg results: List<CheckResult>): List<Message> =
-            results.flatMap { chunk ->
-                chunk.flatMap { result ->
-                    when (result) {
-                        is Warning -> result.warnings
-                        is Failure -> result.errors
-                        else -> emptyList()
-                    }
-                }
-            }
-    }
-}
-
-private fun interface RequirementChecker {
-    fun checked(
-        classDeclaration: KSClassDeclaration,
-        lensableProps: List<KSPropertyDeclaration>
-    ): CheckResult
-}
-
-private class AggregatingRequirementCheckerDecorator(private val checkers: List<RequirementChecker>) :
-    RequirementChecker {
-
-    constructor(vararg checker: RequirementChecker) : this(listOf(*checker))
-
-    fun with(checker: RequirementChecker): AggregatingRequirementCheckerDecorator =
-        AggregatingRequirementCheckerDecorator(checkers + checker)
-
-    override fun checked(
-        classDeclaration: KSClassDeclaration,
-        lensableProps: List<KSPropertyDeclaration>
-    ): CheckResult {
-        val (errors, warnings) = checkers
-            .map { checker -> checker.checked(classDeclaration, lensableProps) }
-            .filterNot { it is CheckResult.Success }
-            .partition { it is CheckResult.Failure }
-        return when {
-            errors.isEmpty() && warnings.isNotEmpty() -> CheckResult.Warning(collectMessages(warnings))
-            errors.isNotEmpty() -> CheckResult.Failure(collectMessages(warnings, errors))
-            else -> CheckResult.Success
-        }
-    }
-}
-
-@JvmInline
-private value class DetermineReservedNames(val get: (List<KSPropertyDeclaration>) -> Set<String>)
-
-private data class LensesPropertyNamesAreAvailable(
-    private val determineReservedNames: DetermineReservedNames,
-    private val companionObject: KSClassDeclaration?
-) : RequirementChecker {
-    override fun checked(
-        classDeclaration: KSClassDeclaration,
-        lensableProps: List<KSPropertyDeclaration>
-    ): CheckResult {
-        val usedNames = companionObject?.let { companion ->
-            (companion.getDeclaredFunctions() + companion.getDeclaredProperties()).map { it.simpleName.getShortName() }
-        }?.toSet() ?: emptySet()
-
-        val conflictingNames = usedNames intersect determineReservedNames.get(lensableProps)
-
-        return if (conflictingNames.isEmpty()) {
-            CheckResult.Success
-        } else {
-            CheckResult.Failure(
-                Message(
-                    Severity.Error,
-                    buildString {
-                        append(
-                            "The companion object of $classDeclaration already defines the following" +
-                                    " functions / properties: "
-                        )
-                        append(" {${conflictingNames.joinToString("; ")}} ")
-                        append(
-                            " -> Those names must not be defined! They are used for the automatic lenses generation. "
-                                    + "Please rename those existing function(s) to bypass this problem!"
-                        )
-                    }
-                )
-            )
-        }
-    }
-
-    companion object {
-        fun forDefaultLensFactories(companionObject: KSClassDeclaration?): LensesPropertyNamesAreAvailable =
-            LensesPropertyNamesAreAvailable(
-                DetermineReservedNames { lensableProps -> lensableProps.map { it.simpleName.getShortName() }.toSet() },
-                companionObject
-            )
-
-        fun forDownTypingLensesInSealedChild(
-            classDeclaration: KSClassDeclaration,
-            resolver: Resolver,
-            companionObject: KSClassDeclaration?
-        ): LensesPropertyNamesAreAvailable =
-            LensesPropertyNamesAreAvailable(
-                DetermineReservedNames { _ ->
-                    val res = classDeclaration.superTypes
-                        .map { "${classDeclaration.packageName.asString()}.$it" }
-                        .mapNotNull { name -> resolver.getClassDeclarationByName(name) }
-                        .filter {
-                            it.isTypeVariant() in setOf(
-                                TypeVariant.SealedDataClass,
-                                TypeVariant.SealedInterface
-                            )
-                        }
-                        .map { it.simpleName.getShortName().lowerCamelCased() }
-                        .toSet()
-                    res
-                },
-                companionObject
-            )
-
-        fun forUpTypingLensesInSealedBase(
-            classDeclaration: KSClassDeclaration,
-            companionObject: KSClassDeclaration?
-        ): LensesPropertyNamesAreAvailable =
-            LensesPropertyNamesAreAvailable(
-                DetermineReservedNames {
-                    val res = classDeclaration.getSealedSubclasses()
-                        .map { it.simpleName.getShortName().lowerCamelCased() }
-                        .toSet()
-                    res
-                },
-                companionObject
-            )
-    }
-}
-
-private val thereAreLenseableProperties = RequirementChecker { classDeclaration, lensableProps ->
-    if (lensableProps.isEmpty()) {
-        CheckResult.Warning(
-            Message(
-                Severity.Warning,
-                buildString {
-                    append("@Lenses annotated data class, sealed class or sealed interface $classDeclaration found,")
-                    if (classDeclaration.isTypeVariant() == TypeVariant.DataClass)
-                        append(" but it has no public properties defined in constructor ")
-                    else
-                        append(" but it has no public properties defined in ist body ")
-                    append("-> can not create any lenses though...")
-                }
-            )
-        )
-    } else {
-        CheckResult.Success
-    }
-}
-
-private val allImplementationsOfSealedTypeAreDataClasses =
-    RequirementChecker { classDeclaration, _ ->
-        val noneDataClassChildren =
-            classDeclaration.getSealedSubclasses().filter { it.isTypeVariant() != TypeVariant.DataClass }.toList()
-        if (noneDataClassChildren.isEmpty()) CheckResult.Success else CheckResult.Failure(
-            Message(
-                Severity.Error,
-                buildString {
-                    appendLine(
-                        "All implementing children of '$classDeclaration' need to be data classes." +
-                                " The following children are no data classes:"
-                    )
-                    noneDataClassChildren.forEach { appendLine("  - ${it.qualifiedName}") }
-                }
-            )
-        )
-    }
-
-private val thereIsAtLeastOneImplementationOfSealedType =
-    RequirementChecker { classDeclaration, _ ->
-        if (classDeclaration.getSealedSubclasses().toList().isNotEmpty()) CheckResult.Success
-        else CheckResult.Warning(
-            Message(
-                Severity.Warning,
-                "There are no subclasses for '$classDeclaration'. -> can not create any lenses though..."
-            )
-        )
-    }
-
-private class CompanionObjectFound(private val companionObject: KSClassDeclaration?) : RequirementChecker {
-    override fun checked(
-        classDeclaration: KSClassDeclaration,
-        lensableProps: List<KSPropertyDeclaration>
-    ): CheckResult =
-        if (companionObject == null) CheckResult.Failure(
-            Message(
-                Severity.Error,
-                "The companion object for $classDeclaration is missing!  Please define it to bypass this error."
-            )
-        )
-        else CheckResult.Success
-}
