+++ lenses-annotation-processor/src/jvmMain/kotlin/dev/fritz2/lens/LowerCamelCased.kt
@@ -1,0 +1,3 @@
+package dev.fritz2.lens
+
+internal fun String.lowerCamelCased() = "${first().lowercase()}${drop(1)}"
